update challenge 4
