+++ javascript-validation-anhduy/package.json
@@ -10,6 +10,7 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.18.2",
+    "express-validation": "^4.1.0",
     "joi": "^17.6.3"
   }
 }
